pyproject.toml
[tool.poetry]
name = "word-selection-assistant"
version = "2.0.0"
description = "AI-powered intelligent text processing assistant with multi-model support"
authors = ["User <user@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"

# GUI Framework
pyqt6 = ">=6.6.0"
pyqt6-webengine = ">=6.6.0"

# AI and API
openai = ">=1.0.0"
aiohttp = ">=3.9.0"
requests = ">=2.31.0"

# System Integration
pywin32 = ">=306"
keyboard = ">=0.13.5"
pyperclip = ">=1.8.2"

# Configuration and Utils
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0.1"
pillow = ">=9.0.0"

# Logging and Development
colorlog = ">=6.10.1"

# Chart Generation (图表生成功能)
matplotlib = ">=3.7.0"
numpy = ">=1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
timeout = 30